Rewrites the plugin to correct errors
Fixes #23 and fixes #15. Rewrites `findTab` to synchronous check if a file exists. If found, the file is copied over to the public patterns folder, or if not found, and empty file is created in its place. Also refactors the JS for improved clarity. This also closes #22 by updating the README.