Adds highlighting language data into snippets
This somewhat addresses #5, but requires that additional fixes be applied to the `patternlab-viewer.js`. More specifically, it requires changing the way the `Prism.highlight` identifies the appropriate language to use. Currently, `patternlab-viewer.js` defaults to using `Prism.languages.html`. In sum, I will also be forking and updating the `styleguidekit-assets-default` package with these changes in order to allow for the snippets language to be passed in via `Panels.add`.